requirements.txt
beautifulsoup4==4.6.0
mechanize==0.3.6
PyYAML==3.12
requests==2.18.4
